:truck: refactor: Chat 관련 소스 도메인으로 이동 (#66)
